test spacio

+++ Yonti/testSpecio.py
@@ -79,7 +79,7 @@ def fingerprint_3D_spatiogram(image, mask):
 
     hist_list = []
     for mask in masks:
-        hist = cv2.calcHist([hsv], [0, 1, 2], mask, bins, [0, 180, 0, 256, 0, 256])
+        hist = (cv2.calcHist([hsv], [0, 1, 2], mask, bins, [0, 180, 0, 256, 0, 256])).astype(np.int32)
         hist_list.append(cv2.normalize(hist, hist).flatten())
     # return np.array(hist_list)
     return hist_list
